requirements.txt
autopep8           1.6.0
beautifulsoup4     4.10.0
certifi            2021.10.8
charset-normalizer 2.0.9
comtypes           1.1.10
idna               3.3

(venv_project1) C:\Users\mayeenulislam\Desktop\python>pip freeze
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
comtypes==1.1.10
idna==3.3
MouseInfo==0.1.3
Pillow==8.4.0
PyAudio @ file:///C:/Users/mayeenulislam/Desktop/python/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
PyAutoGUI==0.9.53
pycodestyle==2.8.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.28
python-decouple==3.5
pyttsx3==2.90
pytweening==1.0.4
pywhatkit==5.2
pywin32==302
requests==2.26.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
toml==0.10.2
urllib3==1.26.7
wikipedia==1.4.0